+++ changelog.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed the `install-bx-module` to include the right helpers.
+
 ## [1.12.1] - 2025-06-17
 
 ### Fixed